Arreglando emisión de factura
