+++ cr_electronic_invoice/__manifest__.py
@@ -11,7 +11,7 @@
 		'''
 		Facturación electronica Costa Rica.
 		''',
-	'depends': ['base', 'account','product','sales_team','l10n_cr_country_codes','account_cancel','point_of_sale'],
+	'depends': ['base', 'account','product','sales_team','l10n_cr_country_codes','account_cancel'],
 	'data': ['data/data.xml',
 	         'data/code.type.product.csv',
 	         'data/identification.type.csv',
