+++ cr_electronic_invoice/models/functions.py
@@ -30,6 +30,7 @@ def get_clave(self, url, tipo_documento, next_number):
     payload['situacion'] = 'normal'
     payload['codigoSeguridad'] = self.company_id.security_code
 
+    _logger.debug(payload)
     response = requests.request("POST", url, data=payload, headers=headers)
     response_json = json.loads(response._content)
     return response_json
@@ -38,7 +39,7 @@ def get_clave(self, url, tipo_documento, next_number):
 def make_xml_invoice(inv, tipo_documento, consecutivo, date, sale_conditions, medio_pago, total_servicio_gravado,
                      total_servicio_exento, total_mercaderia_gravado, total_mercaderia_exento, base_total, lines,
                      tipo_documento_referencia, numero_documento_referencia, fecha_emision_referencia,
-                     codigo_referencia, razon_referencia, url):
+                     codigo_referencia, razon_referencia, url, currency_rate):
     headers = {}
     payload = {}
     # Generar FE payload
@@ -80,23 +81,21 @@ def make_xml_invoice(inv, tipo_documento, consecutivo, date, sale_conditions, me
     payload['plazo_credito'] = ''
     payload['medio_pago'] = medio_pago
     payload['cod_moneda'] = inv.currency_id.name
-    payload['tipo_cambio'] = 1
+    payload['tipo_cambio'] = currency_rate
     payload['total_serv_gravados'] = total_servicio_gravado
     payload['total_serv_exentos'] = total_servicio_exento
     payload['total_merc_gravada'] = total_mercaderia_gravado
     payload['total_merc_exenta'] = total_mercaderia_exento
     payload['total_gravados'] = total_servicio_gravado + total_mercaderia_gravado
     payload['total_exentos'] = total_servicio_exento + total_mercaderia_exento
-    payload[
-        'total_ventas'] = total_servicio_gravado + total_mercaderia_gravado + total_servicio_exento + total_mercaderia_exento
+    payload['total_ventas'] = total_servicio_gravado + total_mercaderia_gravado + total_servicio_exento + total_mercaderia_exento
     payload['total_descuentos'] = round(base_total, 2) - round(inv.amount_untaxed, 2)
-    payload['total_ventas_neta'] = (total_servicio_gravado + total_mercaderia_gravado
-                                    + total_servicio_exento + total_mercaderia_exento) \
-                                   - (base_total - inv.amount_untaxed)
+    payload['total_ventas_neta'] = (total_servicio_gravado + total_mercaderia_gravado + total_servicio_exento + total_mercaderia_exento) - (base_total - inv.amount_untaxed)
     payload['total_impuestos'] = inv.amount_tax
     payload['total_comprobante'] = inv.amount_total
     payload['otros'] = ''
     payload['detalles'] = lines
+
     if tipo_documento == 'NC':
         payload['infoRefeTipoDoc'] = tipo_documento_referencia
         payload['infoRefeNumero'] = numero_documento_referencia
@@ -104,7 +103,9 @@ def make_xml_invoice(inv, tipo_documento, consecutivo, date, sale_conditions, me
         payload['infoRefeCodigo'] = codigo_referencia
         payload['infoRefeRazon'] = razon_referencia
 
+    _logger.debug(payload)
     response = requests.request("POST", url, data=payload, headers=headers)
+    _logger.debug(response)
     response_json = json.loads(response._content)
     return response_json
