+++ cr_electronic_invoice/views/account_journal_views.xml
@@ -6,7 +6,7 @@
 			<field name="model">account.journal</field>
 			<field name="inherit_id" ref="account.view_account_journal_form"/>
 			<field name="arch" type="xml">
-				<xpath expr="//field[@name='refund_sequence_id']" position="after">
+				<xpath expr="//field[@name='type']" position="after">
 					<field name="terminal" attrs="{'invisible':[('type','!=','sale')],'required':[('type','=','sale')]}"/>
 					<field name="sucursal" attrs="{'invisible':[('type','!=','sale')],'required':[('type','=','sale')]}"/>
 				</xpath>
