+++ pos_electronic_invoice/__manifest__.py
@@ -1,0 +1,18 @@
+# -*- coding: utf-8 -*-
+
+{
+	'name': 'Tiquete Electronico Costa Rica',
+	'version': '0.1',
+	'author': 'JackDevelopers',
+	'license': 'OPL-1',
+	'website': 'https://www.jackdevelopers.com',
+	'category': 'pos',
+	'description':
+		'''
+		Tiquete Electronico Costa Rica.
+		''',
+	'depends': ['base', 'cr_electronic_invoice', 'point_of_sale'],
+	'data': [],
+    'qweb': [],
+	'installable': True,
+}
