+++ pos_electronic_invoice/models/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+
+from . import pos_electronic_invoice
